requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
jinja2
assemblyai
google-generativeai